test: add controller genesis test and fixup host test
