+++ modules/apps/27-interchain-accounts/controller/keeper/genesis_test.go
@@ -3,13 +3,32 @@ package keeper_test
 import (
 	"github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/controller/keeper"
 	"github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/controller/types"
+	icacontrollertypes "github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/controller/types"
 	genesistypes "github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/genesis/types"
 	icatypes "github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/types"
+	host "github.com/cosmos/ibc-go/v8/modules/core/24-host"
 	ibctesting "github.com/cosmos/ibc-go/v8/testing"
 )
 
 func (suite *KeeperTestSuite) TestInitGenesis() {
-	suite.SetupTest()
+	ports := []string{"port1", "port2", "port3"}
+
+	testCases := []struct {
+		name     string
+		malleate func()
+	}{
+		{
+			"sucess", func() {},
+		},
+		{
+			"success: capabilities already initialized for first port",
+			func() {
+				capability := suite.chainA.GetSimApp().IBCKeeper.PortKeeper.BindPort(suite.chainA.GetContext(), ports[0])
+				err := suite.chainA.GetSimApp().ICAControllerKeeper.ClaimCapability(suite.chainA.GetContext(), capability, host.PortPath(ports[0]))
+				suite.Require().NoError(err)
+			},
+		},
+	}
 
 	interchainAccAddr := icatypes.GenerateAddress(suite.chainB.GetContext(), ibctesting.FirstConnectionID, TestPortID)
 	genesisState := genesistypes.ControllerGenesisState{
@@ -34,28 +53,47 @@ func (suite *KeeperTestSuite) TestInitGenesis() {
 				AccountAddress: interchainAccAddr.String(),
 			},
 		},
-		Ports: []string{TestPortID},
+		Ports: ports,
+	}
+	for _, tc := range testCases {
+		tc := tc
+
+		suite.Run(tc.name, func() {
+			suite.SetupTest()
+
+			tc.malleate()
+
+			keeper.InitGenesis(suite.chainA.GetContext(), suite.chainA.GetSimApp().ICAControllerKeeper, genesisState)
+
+			channelID, found := suite.chainA.GetSimApp().ICAControllerKeeper.GetActiveChannelID(suite.chainA.GetContext(), ibctesting.FirstConnectionID, TestPortID)
+			suite.Require().True(found)
+			suite.Require().Equal(ibctesting.FirstChannelID, channelID)
+
+			isMiddlewareEnabled := suite.chainA.GetSimApp().ICAControllerKeeper.IsMiddlewareEnabled(suite.chainA.GetContext(), TestPortID, ibctesting.FirstConnectionID)
+			suite.Require().True(isMiddlewareEnabled)
+
+			isMiddlewareDisabled := suite.chainA.GetSimApp().ICAControllerKeeper.IsMiddlewareDisabled(suite.chainA.GetContext(), "test-port-1", "connection-1")
+			suite.Require().True(isMiddlewareDisabled)
+
+			accountAdrr, found := suite.chainA.GetSimApp().ICAControllerKeeper.GetInterchainAccountAddress(suite.chainA.GetContext(), ibctesting.FirstConnectionID, TestPortID)
+			suite.Require().True(found)
+			suite.Require().Equal(interchainAccAddr.String(), accountAdrr)
+
+			expParams := types.NewParams(false)
+			params := suite.chainA.GetSimApp().ICAControllerKeeper.GetParams(suite.chainA.GetContext())
+			suite.Require().Equal(expParams, params)
+
+			for _, port := range ports {
+				store := suite.chainA.GetContext().KVStore(suite.chainA.GetSimApp().GetKey(icacontrollertypes.StoreKey))
+				suite.Require().True(store.Has(icatypes.KeyPort(port)))
+
+				capability, found := suite.chainA.GetSimApp().ScopedICAControllerKeeper.GetCapability(suite.chainA.GetContext(), host.PortPath(port))
+				suite.Require().True(found)
+				suite.Require().NotNil(capability)
+			}
+		})
 	}
 
-	keeper.InitGenesis(suite.chainA.GetContext(), suite.chainA.GetSimApp().ICAControllerKeeper, genesisState)
-
-	channelID, found := suite.chainA.GetSimApp().ICAControllerKeeper.GetActiveChannelID(suite.chainA.GetContext(), ibctesting.FirstConnectionID, TestPortID)
-	suite.Require().True(found)
-	suite.Require().Equal(ibctesting.FirstChannelID, channelID)
-
-	isMiddlewareEnabled := suite.chainA.GetSimApp().ICAControllerKeeper.IsMiddlewareEnabled(suite.chainA.GetContext(), TestPortID, ibctesting.FirstConnectionID)
-	suite.Require().True(isMiddlewareEnabled)
-
-	isMiddlewareDisabled := suite.chainA.GetSimApp().ICAControllerKeeper.IsMiddlewareDisabled(suite.chainA.GetContext(), "test-port-1", "connection-1")
-	suite.Require().True(isMiddlewareDisabled)
-
-	accountAdrr, found := suite.chainA.GetSimApp().ICAControllerKeeper.GetInterchainAccountAddress(suite.chainA.GetContext(), ibctesting.FirstConnectionID, TestPortID)
-	suite.Require().True(found)
-	suite.Require().Equal(interchainAccAddr.String(), accountAdrr)
-
-	expParams := types.NewParams(false)
-	params := suite.chainA.GetSimApp().ICAControllerKeeper.GetParams(suite.chainA.GetContext())
-	suite.Require().Equal(expParams, params)
 }
 
 func (suite *KeeperTestSuite) TestExportGenesis() {
